moved rtioclkmult and serdes to separate file
